refactor: SwapButton
